Refactor password change: move to own page
