+++ frontend/src/pages/Profile.js
@@ -5,11 +5,7 @@ import Error from '../components/Error'
 import Main from '../components/Main'
 import { useCallback, useEffect, useState } from 'react'
 import { useAuth } from '../auth/AuthProvider'
-import {
-  getUserAddressList,
-  putEditUsername,
-  updatePassword,
-} from '../services/api-service'
+import { getUserAddressList, putEditUsername } from '../services/api-service'
 import { useHistory, useParams } from 'react-router-dom'
 import Button from '../components/Button'
 import Addresses from '../components/Addresses'
@@ -21,22 +17,14 @@ import MainBottom from '../components/MainBottom'
 import styled from 'styled-components/macro'
 import save from '../images/save-32.png'
 
-const initialPW = {
-  oldPassword: '',
-  newPassword: '',
-  retypePassword: '',
-}
 export default function Profile() {
   const { mode, id } = useParams()
-  const { user, token, login, logout } = useAuth()
+  const { user, token, logout } = useAuth()
   const history = useHistory()
   const [loading, setLoading] = useState(false)
   const [error, setError] = useState()
   const [addressList, setAddressList] = useState([])
   const [username, setUsername] = useState(user.username)
-  const [passwords, setPasswords] = useState(initialPW)
-
-  const [viewPW, setViewPW] = useState(false)
 
   const loadDataOnlyOnce = useCallback(() => {
     setLoading(true)
@@ -76,32 +64,9 @@ export default function Profile() {
       .finally(() => setLoading(false))
   }
 
-  const togglePasswordChange = () => {
-    setViewPW(!viewPW)
+  const handleChangePassword = () => {
+    history.push(`/password/change`)
   }
-
-  const handleChange = event => {
-    setPasswords({ ...passwords, [event.target.name]: event.target.value })
-  }
-
-  const submitNewPassword = () => {
-    setLoading(true)
-    setError()
-    updatePassword(token, passwords)
-      .then(() => {
-        login({ username: user.username, password: passwords.newPassword })
-        setPasswords(initialPW)
-        setViewPW(false)
-      })
-      .catch(setError)
-      .finally(() => {
-        setLoading(false)
-      })
-  }
-
-  const passwordMatch =
-    passwords.newPassword.length &&
-    passwords.newPassword === passwords.retypePassword
 
   return (
     <Page>
@@ -130,41 +95,11 @@ export default function Profile() {
               </Button>
             </User>
             <div>
-              <Button type="button" onClick={togglePasswordChange}>
+              <Button type="button" onClick={handleChangePassword}>
                 Passwort ändern
               </Button>
             </div>
-            {viewPW && (
-              <Password>
-                <TextField
-                  name="oldPassword"
-                  value={passwords.oldPassword}
-                  onChange={handleChange}
-                  title="Altes Passwort"
-                  type="password"
-                  autocomplete="off"
-                />
-                <TextField
-                  name="newPassword"
-                  value={passwords.newPassword}
-                  onChange={handleChange}
-                  title="Neues Passwort (wiederholen)"
-                  type="password"
-                  autocomplete="off"
-                />
-                <TextField
-                  name="retypePassword"
-                  value={passwords.retypePassword}
-                  onChange={handleChange}
-                  title="Neues Passwort"
-                  type="password"
-                  autocomplete="off"
-                />
-                <Button disabled={!passwordMatch} onClick={submitNewPassword}>
-                  speichern
-                </Button>
-              </Password>
-            )}
+
             <Addresses
               user={user}
               mode={mode}
@@ -206,4 +141,3 @@ const User = styled.form`
     height: min-content;
   }
 `
-const Password = styled.form``
